requirements.txt
psycopg==3.1.19
SQLAlchemy==2.0.31
python-decouple==3.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
streamlit==1.36.0
